move quaternion yaw back
