+++ python-sdk/nuscenes/utils/geometry_utils.py
@@ -108,21 +108,3 @@ def transform_matrix(translation: np.ndarray = np.array([0, 0, 0]),
         tm[:3, 3] = np.transpose(np.array(translation))
 
     return tm
-
-
-def quaternion_yaw(q: Quaternion) -> float:
-    """
-    Calculate the yaw angle from a quaternion.
-    Note that this only works for a quaternion that represents a box in lidar or global coordinate frame.
-    It does not work for a box in the camera frame.
-    :param q: Quaternion of interest.
-    :return: Yaw angle in radians.
-    """
-
-    # Project into xy plane.
-    v = np.dot(q.rotation_matrix, np.array([1, 0, 0]))
-
-    # Measure yaw using arctan.
-    yaw = np.arctan2(v[1], v[0])
-
-    return yaw
