+++ python-sdk/nuscenes/utils/tests/test_geometry_utils.py
@@ -7,7 +7,7 @@ import unittest
 import numpy as np
 from pyquaternion import Quaternion
 
-from nuscenes.utils.geometry_utils import quaternion_yaw
+from nuscenes.eval.detection.utils import quaternion_yaw
 
 
 class TestGeometryUtils(unittest.TestCase):
